pb 5 800 200 200

+++ check_end_sim.c
@@ -6,7 +6,7 @@
 /*   By: dchheang <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/14 09:33:29 by dchheang          #+#    #+#             */
-/*   Updated: 2022/02/08 13:15:17 by dchheang         ###   ########.fr       */
+/*   Updated: 2022/02/08 14:02:21 by dchheang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ int	check_death(t_philo *philo, t_info *info)
 
 	ret = 0;
 	pthread_mutex_lock(&info->death_mutex);
-	if (get_timediff(philo->time_last_meal) >= (unsigned long)info->time_to_die)
+	if (get_timediff(philo->time_last_meal) >= (long long)info->time_to_die)
 	{
 		info->end_sim = 1;
-		printf("%lu % d %s\n", get_timediff(info->time_start), philo->id, "died");
+		printf("%lld % d %s\n", get_timediff(info->time_start), philo->id, "died");
 		ret = 1;
 	}
 	pthread_mutex_unlock(&info->death_mutex);
